[REFACTOR] my 컴포넌트 세부 도메인 분리
