+++ src/containers/my/profile/ProfileForm.tsx
@@ -5,7 +5,7 @@ import { useMemo, useState } from "react";
 
 import Header from "@/components/common/Header";
 import TextArea from "@/components/common/TextArea";
-import ChipGroup from "@/containers/my/ChipGroup";
+import ChipGroup from "@/containers/my/profile/ChipGroup";
 import { JOB_OPTIONS, STATUS_OPTIONS } from "@/constants/my";
 import { NICKNAME_REGEX } from "@/constants/regex";
 import { getNicknameError } from "@/lib/utils/validation";
